% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_postprocess.R
\name{align.MetMatchResults}
\alias{align.MetMatchResults}
\title{align.MetMatchResults}
\usage{
align.MetMatchResults(wd, cutoff = 0.1, N = 1, maxRtDiff = 60)
}
\arguments{
\item{wd}{The working directory containing the metmatch results}

\item{cutoff}{A minimum score.  Must be >=0;}

\item{N}{Minimum number of samples in which an alinged FFM must be present.}

\item{maxRtDiff}{maximum retention time deviation (sec) in which the formulae must be present between samples}
}
\description{
Postprocesses groups of MetMatch results by formula and retention time.
}

