% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{build.DB.emass}
\alias{build.DB.emass}
\title{build.DB.emass}
\usage{
build.DB.emass(k, formula)
}
\arguments{
\item{k}{number of isotopes}

\item{formula}{The input list of formulae}
}
\value{
The target feature library of emass predicted isotopic mass and intensities in DB object form.
}
\description{
Builds the mass and intensity database using emass.
}

