% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{clean.formulae}
\alias{clean.formulae}
\title{clean.formulae}
\usage{
clean.formulae(formulae)
}
\arguments{
\item{formulae}{}
}
\value{
the character vector of clean formulae
}
\description{
Cleans Formulae of unwanted entries, and trims the list down to unique entries.
}

