% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{compute.delta.int}
\alias{compute.delta.int}
\title{Compute Isotope Delta intensities}
\usage{
compute.delta.int(int.Q, int.DB)
}
\arguments{
\item{int.Q}{The query intensity}

\item{int.DB}{The database of intensities}
}
\value{
Array of delta intensities
}
\description{
Computes the delta intensities for a single query isotope intensities compared to all intensities in the database
}
\keyword{internal}

