% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{compute.delta.mass}
\alias{compute.delta.mass}
\title{Compute Isotope Delta Masses}
\usage{
compute.delta.mass(mass.Q, mass.DB)
}
\arguments{
\item{mass.Q}{query mass}

\item{mass.DB}{Database of masses}
}
\value{
Array of delta masses
}
\description{
Computes the delta masses for a single query isotope masses compared to all masses in the database
}
\keyword{internal}

