% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{compute.Deltas}
\alias{compute.Deltas}
\title{compute.Deltas}
\usage{
compute.Deltas(query.isotope.data, q.data, db.data, mass.tol, rel.int.tol,
  score.type, Mono = FALSE)
}
\arguments{
\item{rel.int.tol}{}
}
\value{
lists of data frames containing the computed delta values, and unscored peaks.
}
\description{
Computes the delta scores for each isotopic peak.
}

