% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{compute.FDR}
\alias{compute.FDR}
\title{compute.FDR}
\usage{
compute.FDR(dec.scored.data, dec.db.data, score.method)
}
\arguments{
\item{dec.db.data}{}
}
\value{
Lists of scored and unscored data with FDR
}
\description{
computed the FDR when using the decoy option in  metmatch.
}

