% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{compute.model.Scores}
\alias{compute.model.Scores}
\title{compute.model.Scores}
\usage{
compute.model.Scores(delta.data, delta.models, score.method, score.type)
}
\arguments{
\item{score.type}{}
}
\value{
lists of scored and unscored data.
}
\description{
Performs the log liklihood scoring
}

