% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{compute.thresholds}
\alias{compute.thresholds}
\title{compute.thresholds}
\usage{
compute.thresholds(dec.scored.data)
}
\arguments{
\item{dec.scored.data}{}
}
\description{
compute.thresholds
}
\keyword{internal}

