% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{convert.camera.file}
\alias{convert.camera.file}
\title{Load the CAMERA output file into a list of query objects}
\usage{
convert.camera.file(fh, N)
}
\arguments{
\item{fh}{file handle}

\item{N}{N samples}
}
\value{
Q class
}
\description{
Load the CAMERA output file into a list of query objects
}

