% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{convertIonToNeutralMass}
\alias{convertIonToNeutralMass}
\title{convertIonsToNeutralMass}
\usage{
convertIonToNeutralMass(ionType, mz)
}
\arguments{
\item{ionType}{the input ion type}

\item{mz}{the input mz}
}
\value{
the neutral mass
}
\description{
Converts ion mz to M neutral charge mass.
}
\keyword{internal}

