% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{fix.camera.ions}
\alias{fix.camera.ions}
\title{Fix CAMERA (xcms) annotations}
\usage{
fix.camera.ions(ions)
}
\arguments{
\item{ions}{the input ion type}
}
\description{
Helper function for loading CAMERA annotations
}
\keyword{internal}

