% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{generate.decoy.DB}
\alias{generate.decoy.DB}
\title{generate.decoy.DB}
\usage{
generate.decoy.DB(DB, jM, jI, k)
}
\arguments{
\item{DB}{The input database}

\item{jM}{Jitter masses (ppm)}

\item{jI}{Jitter intensities (0-1)}

\item{k}{k number of decoys produced per input spectra}
}
\value{
Returns the decoy feature library in DB object form.
}
\description{
Produce a decoy feature library by computing around mass (ppm) and %intensity.
}
\keyword{internal}

