% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{generate.decoy.spectra}
\alias{generate.decoy.spectra}
\title{generate.decoy.spectra}
\usage{
generate.decoy.spectra(DB, jM, dI, k)
}
\arguments{
\item{DB}{The input target feature library}

\item{jM}{Jitter the mass in ppm}

\item{dI}{Second best delta intensities}

\item{k}{Number of decoys produced per input spectra}
}
\value{
Returns the decoy feature library in DB object form.
}
\description{
Produce a decoy database by computing around mass (ppm) and %intensity.
}

