% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{get.dataframe}
\alias{get.dataframe}
\title{Format dataframe of masses or Intensities}
\usage{
get.dataframe(X, n, type = "mass")
}
\arguments{
\item{X}{Input of query masses or intensites}

\item{n}{Number of isotope peaks allowed}

\item{type}{"mass" or "I" intensity}
}
\value{
Array of mass or intensity vectors
}
\description{
Format dataframe of masses or Intensities
}
\keyword{internal}

