% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{get.DB.dataframe}
\alias{get.DB.dataframe}
\title{get.DB.dataframe}
\usage{
get.DB.dataframe(DB, n)
}
\arguments{
\item{DB}{The database (list of DB objects)}

\item{n}{Limit to n number of peaks in the output for each isotope}
}
\value{
The database in dataframe format.
}
\description{
A helper function to convert list of DB objects into a dataframe.
}

