% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{get.query.Isotope.data}
\alias{get.query.Isotope.data}
\title{get.query.Isotope.data}
\usage{
get.query.Isotope.data(q.data)
}
\arguments{
\item{q.data}{}
}
\value{
list of experimental isotopic peak  data.
}
\description{
Organizes lists of experimental isotope data. Internal function.
}
\keyword{internal}

