% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wrappers.R
\name{getIsotopes}
\alias{getIsotopes}
\title{Annotate feature isotopes.}
\usage{
getIsotopes(xs, fh, ppm = 5, mzabs = 0.02, sigma = 3)
}
\arguments{
\item{xs}{The XCMS xset}

\item{ppm}{CAMERA mz tolerance window in ppm}

\item{mzabs}{CAMERA mz deviation in mz}

\item{sigma}{sigma for CAMERA groupFWHM function}
}
\description{
Runs the CAMERA pipeline.
}

