% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_postprocess.R
\name{getMetMatchResults}
\alias{getMetMatchResults}
\title{getMetMatchResults}
\usage{
getMetMatchResults(metFiles, cutoff = 0)
}
\arguments{
\item{cutoff}{}
}
\value{
Dataframe of values used for alignent by rt and formulae.
}
\description{
Extracts a data frame of results from the group of metmatch results.
}

