% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_postprocess.R
\name{groupIonsByFilename}
\alias{groupIonsByFilename}
\title{groupIonsByFilename}
\usage{
groupIonsByFilename(g.ions, metFiles)
}
\arguments{
\item{metFiles}{}
}
\description{
Helper function for prostprocessing
}
\keyword{internal}

