% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_postprocess.R
\name{groupIonsByRT}
\alias{groupIonsByRT}
\title{groupIonsByRT}
\usage{
groupIonsByRT(results, maxRtDiff = 60)
}
\arguments{
\item{maxRtDiff}{}
}
\value{
List of aggregated data, aggregated by formula and RT.
}
\description{
Groups memtmatch results into data chunks organized by formula and separated by a retention time clutering method.
}
\keyword{internal}

