% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{laplace.unif.EM}
\alias{laplace.unif.EM}
\title{Mixture model for Laplace Uniform distribution}
\usage{
laplace.unif.EM(x, tol = 1e-05, maxiter = 100)
}
\arguments{
\item{x}{Input array of values of X}

\item{tol}{Tolerance for minimizaton}

\item{maxiter}{Maximum number of iterations}
}
\value{
List of parameters
}
\description{
Performs expectation-maximization to fit the Laplace-Uniform distributions given input set of X values.
}
\keyword{internal}

