% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{logdlaplace}
\alias{logdlaplace}
\title{log Laplace Distribution}
\usage{
logdlaplace(x,m,b)
}
\arguments{
\item{x}{Single value or array of values}

\item{m}{mu}

\item{b}{scale}
}
\value{
y value of Laplace distribution
}
\description{
Given a set of x values, mean and scale parameter, returns a log Laplace distribution.
}
\keyword{internal}

