% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{massTol}
\alias{massTol}
\title{Compute Mass Tolerance}
\usage{
massTol(M, ppm)
}
\arguments{
\item{M}{Input neutral mass}

\item{ppm}{Desired ppm tolerance}
}
\value{
Mass tolerance
}
\description{
Computes the mass tolerance window for given mass and desired ppm
}
\keyword{internal}

