% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metmatch.R
\name{metmatch}
\alias{metmatch}
\title{MetMatch, the molecular formula scoring algorithm with local error rate estimation}
\usage{
metmatch(Q.data, DB, mass.tol = 50, rel.int.tol = 0.5, Mono = FALSE,
  score.method = "best", run.type = "nodecoy", score.type = "both",
  out.file.name)
}
\arguments{
\item{Q.data}{The Query data, a list of query objects}

\item{DB}{The search database.  When using a decoy database, be sure that run.type option is set to "decoy"}

\item{mass.tol}{Mass tolerance in part per million (ppm)}

\item{score.method}{Scoring method, options are "best" (default) or "delta"}

\item{run.type}{Options are "nodecoy" (default) or "decoy".  Choose decoy when using a decoy database.}

\item{score.type}{Score type, options are "both", "mass" or "intensity"}

\item{int.rel.tol}{Relative intensity tolerance (0-1)}
}
\value{
List of query-results objects.
}
\description{
MetMatch, the molecular formula scoring algorithm with local error rate estimation
}

