% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{model.Deltas}
\alias{model.Deltas}
\title{model.Deltas.}
\usage{
model.Deltas(delta.data)
}
\arguments{
\item{delta.data}{}
}
\value{
List of model parameters
}
\description{
Returns the Laplace-uniform model parameters after EM.
}

