% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{plotHeatMap}
\usage{
plotHeatMap(mergedData, which = "intensity", name)
}
\arguments{
\item{Dataframe}{from align.MetMatchResults}
}
\value{
Writes a png file of the heatmap
}
\description{
plotHeatMap
}

