% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotScores}
\alias{plotScores}
\title{plotScores}
\usage{
plotScores(scored.data, run.type = "decoy")
}
\arguments{
\item{run.type}{run type}

\item{data}{input data}

\item{score.method}{score method}
}
\description{
plotScores
}

