% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{ppmErr}
\alias{ppmErr}
\title{ppmErr}
\usage{
ppmErr(M, ppmTol)
}
\arguments{
\item{M}{the input mass}

\item{ppmTol}{the ppm tolerance window}
}
\value{
the absolute ppm tolerance
}
\description{
Compute the ppmErr for a given mass and expected ppm tolerance
}
\keyword{internal}

