% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{read.DB}
\alias{read.DB}
\title{read.DB}
\usage{
read.DB(file)
}
\arguments{
\item{file}{}
}
\value{
Returns list of DB objects.
}
\description{
Reads the Target Feature Library from a tab delimited file.
}

