% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wrappers.R
\name{run.CAMERA}
\alias{run.CAMERA}
\title{run.CAMERA}
\usage{
run.CAMERA(wd, method = "massifquant", prefilter = c(3, 500),
  peakwidth = c(3, 45), snthresh = 1.5, withWave = 1, save = FALSE, ...)
}
\arguments{
\item{wd}{Working directory containing the raw mzxml files}

\item{save}{Saves the XCMS set as a RDA file.}

\item{...}{}
}
\description{
Wrapper function to run the XCMS/CAMERA feature finding software.This is a wrapper for a single or group of mzXML files in a directory.
}

