% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{run.emass}
\alias{run.emass}
\title{run.emass}
\usage{
run.emass(formula, maxPks = 6)
}
\arguments{
\item{formula}{The input Lewis formula}

\item{maxPks}{Maximum number of desired peaks}
}
\value{
Returns isotopic (mass and intensity) values
}
\description{
Runs Emass for prediciton of isotopic abundances
}
\keyword{internal}

