% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wrappers.R
\name{run.metmatch}
\alias{run.metmatch}
\title{run.metmatch}
\usage{
run.metmatch(wd, DB, save = TRUE, score.method = "best",
  run.type = "nodecoy", score.type = "both")
}
\arguments{
\item{...}{}
}
\description{
Wrapper function for searching all isotope files in a directory.
}

