% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{write.DB}
\alias{write.DB}
\title{write.DB}
\usage{
write.DB(file, DB)
}
\arguments{
\item{file}{File name to save as}

\item{DB}{The list of DB objects.}
}
\description{
Writes the Target Feature Library in tab deliminted form.
}

